+++ CHANGELOG.md
@@ -44,6 +44,10 @@ Please add one entry in this file for each change in Yarn's behavior. Use the sa
 
   [#6568](https://github.com/yarnpkg/yarn/pull/6568) - [**Hidde Boomsma**](https://github.com/hboomsma)
 
+- Fixes `require.resolve` when used together with the `paths` option
+
+  [#6565](https://github.com/yarnpkg/yarn/pull/6565) - [**Maël Nison**](https://twitter.com/arcanis)
+
 ## 1.12.0
 
 - Adds initial support for PnP on Windows